+++ packages/browser-sdk/CHANGELOG.md
@@ -20,6 +20,15 @@
   - @phantom/chains@1.0.0-beta.0
   - @phantom/base64url@1.0.0-beta.0
 
+## 0.3.9
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/embedded-provider-core@0.1.10
+  - @phantom/client@0.1.11
+
 ## 0.3.8
 
 ### Patch Changes