+++ packages/embedded-provider-core/CHANGELOG.md
@@ -19,6 +19,16 @@
   - @phantom/api-key-stamper@1.0.0-beta.0
   - @phantom/base64url@1.0.0-beta.0
 
+## 0.1.10
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/client@0.1.11
+
+## 0.1.9
+
 ## 0.1.8
 
 ### Patch Changes