+++ packages/react-native-sdk/CHANGELOG.md
@@ -21,9 +21,20 @@
   - @phantom/base64url@1.0.0-beta.0
   - @phantom/crypto@1.0.0-beta.0
 
+## 0.1.8
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/embedded-provider-core@0.1.10
+  - @phantom/client@0.1.11
+
 ## 0.1.7
 
 ### Patch Changes
+
+- Updated dependencies
   - @phantom/embedded-provider-core@0.1.9
   - @phantom/sdk-types@0.1.5
   - @phantom/client@0.1.10