+++ packages/react-sdk/CHANGELOG.md
@@ -15,6 +15,14 @@
   - @phantom/constants@1.0.0-beta.0
   - @phantom/chains@1.0.0-beta.0
 
+## 0.3.9
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/browser-sdk@0.3.9
+
 ## 0.3.8
 
 ### Patch Changes