+++ packages/server-sdk/CHANGELOG.md
@@ -16,6 +16,14 @@
   - @phantom/api-key-stamper@1.0.0-beta.0
   - @phantom/base64url@1.0.0-beta.0
 
+## 0.2.5
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/client@0.1.11
+
 ## 0.2.4
 
 ### Patch Changes